Refactored Models into separate namespaces/directories
