+++ app/BusinessLogicLayer/CurrentQuestionnaireProvider.php
@@ -4,7 +4,7 @@
 namespace App\BusinessLogicLayer;
 
 
-use App\Models\Questionnaire;
+use App\Models\Questionnaire\Questionnaire;
 use App\Repository\Questionnaire\QuestionnaireRepository;
 use App\Repository\Questionnaire\Responses\QuestionnaireResponseRepository;
