+++ app/Models/CrowdSourcingProject/CrowdSourcingProjectStatusLkp.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Models\CrowdSourcingProject;
+
+
+use Illuminate\Database\Eloquent\Model;
+
+class CrowdSourcingProjectStatusLkp extends Model
+{
+    protected $table = 'crowd_sourcing_project_statuses_lkp';
+}
