+++ app/Models/Questionnaire/Statistics/QuestionnaireStatisticsPageVisibilityLkp.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Questionnaire\Statistics;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
@@ -12,8 +12,8 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  * @property string $title
  * @property string $description
  */
-class QuestionnaireStatisticsPageVisibilityLkp extends Model
-{
+class QuestionnaireStatisticsPageVisibilityLkp extends Model {
     use SoftDeletes;
+
     protected $table = 'questionnaire_statistics_page_visibility_lkp';
 }
