+++ app/Repository/Questionnaire/Statistics/QuestionnaireStatisticsPageVisibilityLkpRepository.php
@@ -4,7 +4,7 @@
 namespace App\Repository\Questionnaire\Statistics;
 
 
-use App\Models\QuestionnaireStatisticsPageVisibilityLkp;
+use App\Models\Questionnaire\Statistics\QuestionnaireStatisticsPageVisibilityLkp;
 use App\Repository\Repository;
 
 class QuestionnaireStatisticsPageVisibilityLkpRepository extends Repository {
